Rename service to match table name
