+++ db.js
@@ -1,7 +1,7 @@
 const Airtable = require('airtable');
 const base = new Airtable({ apiKey: process.env.AIRTABLE_KEY }).base(process.env.AIRTABLE_BASE_ID);
-const things = base('Inventory');
+const inventory = base('Inventory');
 
 module.exports = {
-    things
+    inventory
 };
